Refactor with argparse_helpers
